// src/lib/types/database.types.ts
export type Database = {
	public: {
		Tables: {
			customer: {
				Row: {
					customer_id: string;
					email: string | null;
					image_url: string | null;
					is_archived: boolean;
					name: string;
					tax_registration_number: string | null;
				};
				Insert: {
					customer_id?: string;
					email?: string | null;
					image_url?: string | null;
					is_archived?: boolean;
					name: string;
					tax_registration_number?: string | null;
				};
				Update: {
					customer_id?: string;
					email?: string | null;
					image_url?: string | null;
					is_archived?: boolean;
					name?: string;
					tax_registration_number?: string | null;
				};
				Relationships: [];
			};
// ...
			invoice: {
				Row: {
					created_at: string;
					customer_id: string;
					invoice_date: string;
					invoice_id: string;
					invoice_number: string;
					is_void: boolean;
					notes: string | null;
					paid_amount: number;
					payment_status: Database['public']['Enums']['invoice_payment_status'];
				};
				Insert: {
					created_at?: string;
					customer_id: string;
					invoice_date: string;
					invoice_id?: string;
					invoice_number: string;
					is_void?: boolean;
					notes?: string | null;
					paid_amount?: number;
					payment_status?: Database['public']['Enums']['invoice_payment_status'];
				};
				Update: {
					created_at?: string;
					customer_id?: string;
					invoice_date?: string;
					invoice_id?: string;
					invoice_number?: string;
					is_void?: boolean;
					notes?: string | null;
					paid_amount?: number;
					payment_status?: Database['public']['Enums']['invoice_payment_status'];
				};
				Relationships: [
					{
						foreignKeyName: 'invoice_customer_id_fkey';
						columns: ['customer_id'];
						isOneToOne: false;
						referencedRelation: 'customer';
						referencedColumns: ['customer_id'];
					}
				];
			};
// ...
			product: {
				Row: {
					category_id: string | null;
					is_archived: boolean;
					name: string;
					product_id: string;
					sales_tax_id: string | null;
					unit_price: number | null;
				};
				Insert: {
					category_id?: string | null;
					is_archived?: boolean;
					name: string;
					product_id?: string;
					sales_tax_id?: string | null;
					unit_price?: number | null;
				};
				Update: {
					category_id?: string | null;
					is_archived?: boolean;
					name?: string;
					product_id?: string;
					sales_tax_id?: string | null;
					unit_price?: number | null;
				};
				Relationships: [
					{
						foreignKeyName: 'product_category_id_fkey';
						columns: ['category_id'];
						isOneToOne: false;
						referencedRelation: 'invoice_financial_report';
						referencedColumns: ['product_category_id'];
					},
					{
						foreignKeyName: 'product_category_id_fkey';
						columns: ['category_id'];
						isOneToOne: false;
						referencedRelation: 'product_category';
						referencedColumns: ['category_id'];
					},
					{
						foreignKeyName: 'product_sales_tax_id_fkey';
						columns: ['sales_tax_id'];
						isOneToOne: false;
						referencedRelation: 'sales_tax';
						referencedColumns: ['sales_tax_id'];
					}
				];
			};
			product_category: {
				Row: {
					category_id: string;
					name: string;
					position: number;
				};
				Insert: {
					category_id?: string;
					name: string;
					position: number;
				};
				Update: {
					category_id?: string;
					name?: string;
					position?: number;
				};
				Relationships: [];
			};
			product_inventory: {
				Row: {
					allow_sales_when_out_of_stock: boolean;
					low_stock_threshold: number;
					product_id: string;
					stock: number;
				};
				Insert: {
					allow_sales_when_out_of_stock: boolean;
					low_stock_threshold: number;
					product_id: string;
					stock: number;
				};
				Update: {
					allow_sales_when_out_of_stock?: boolean;
					low_stock_threshold?: number;
					product_id?: string;
					stock?: number;
				};
				Relationships: [
					{
						foreignKeyName: 'product_inventory_product_id_fkey';
						columns: ['product_id'];
						isOneToOne: true;
						referencedRelation: 'product';
						referencedColumns: ['product_id'];
					}
				];
			};
			sales_tax: {
				Row: {
					abbreviation: string;
					is_archived: boolean;
					is_recoverable: boolean;
					name: string;
					rate: number;
					sales_tax_id: string;
				};
				Insert: {
					abbreviation: string;
					is_archived?: boolean;
					is_recoverable: boolean;
					name: string;
					rate: number;
					sales_tax_id?: string;
				};
				Update: {
					abbreviation?: string;
					is_archived?: boolean;
					is_recoverable?: boolean;
					name?: string;
					rate?: number;
					sales_tax_id?: string;
				};
				Relationships: [];
			};
// ...
		};
// ...
	};
};

type PublicSchema = Database[Extract<keyof Database, 'public'>];
// ...
export type Enums<
	PublicEnumNameOrOptions extends keyof PublicSchema['Enums'] | { schema: keyof Database },
	EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
		? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
		: never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
	? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
	: PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
		? PublicSchema['Enums'][PublicEnumNameOrOptions]
		: never;